Added calculating MD5 before upload to be sure existing file is the same to silently ignore. Added regular check during upload to check if file with the same path already exists. Added Upload state to show what is happing with upload like uploading, finishing upload or waiting in the queue. Fixed Cache foder cleaning.

+++ amazon-clouddrive-dokan/SmallFilesCache.cs
@@ -231,19 +231,22 @@
             var failed = 0;
             try
             {
-                foreach (var file in Directory.GetFiles(oldcachePath, "*", SearchOption.AllDirectories).ToList())
+                if (Directory.Exists(oldcachePath))
                 {
-                    try
+                    foreach (var file in Directory.GetFiles(oldcachePath, "*", SearchOption.AllDirectories).ToList())
                     {
-                        File.Delete(file);
-                    }
-                    catch (IOException)
-                    {
-                        failed++;
+                        try
+                        {
+                            File.Delete(file);
+                        }
+                        catch (IOException)
+                        {
+                            failed++;
+                        }
                     }
                 }
             }
-            catch (FileNotFoundException)
+            catch (Exception e) when (e is FileNotFoundException || e is DirectoryNotFoundException)
             {
                 Log.Warn("Old cache folder was empty");
             }
